refactor(sprints): name the feature directory the way the product does
