+++ apps/web/src/app/(app)/sprints/loading.tsx
@@ -1,4 +1,4 @@
-import { CycleSkeleton } from '@/features/cycles/cycle-skeleton.tsx';
+import { CycleSkeleton } from '@/features/sprints/cycle-skeleton.tsx';
 
 export default function CyclesLoading() {
   return <CycleSkeleton panels={2} />;
