+++ apps/web/src/app/(app)/team/[key]/sprint/active/loading.tsx
@@ -1,4 +1,4 @@
-import { CycleSkeleton } from '@/features/cycles/cycle-skeleton.tsx';
+import { CycleSkeleton } from '@/features/sprints/cycle-skeleton.tsx';
 
 export default function ActiveCycleLoading() {
   return <CycleSkeleton panels={1} />;
